Hardcoded references to xalan and xmlserializer in classpath.  Those 2 jars MUST appear before netcdf in the classpath, or this error appears in the console: net.sf.saxon.trans.DynamicError: DOMSource cannot be processed: check that saxon8-dom.jar is on the classpath
This happens because something in java is loading the first
TransformerFactory implementation it can find in the classpath.  Netcdf
provides an incomplete one that generates this exception when used,
xalan provides a full, working one.  It's not enough to load core/lib/*
in the classpath - the expanded order of jars is non-deterministic, and
sometimes places netcdf ahead of xalan.